refactor: improve module boundaries
