+++ .env.local
@@ -1,7 +1,7 @@
 BASE_URL=http://localhost:3000
 
 # Google Analytics Config - commented out to disable tracking in dev environment
-GA_MEASUREMENT_ID=UA-169820463-2
+# GA_MEASUREMENT_ID=UA-169820463-2
 
 # Auth0 Config - dedicated development tenant, not used in production
 AUTH0_DOMAIN=gomatcha-dev.eu.auth0.com
@@ -12,7 +12,7 @@ AUTH0_COOKIE_SECRET=OTvWrsWG2MGsFDGXmTPQXajqTK8t8rzPoz2xkhlYYljpCh9RD9iovJTI6AMe
 # MailChimp Config - we only have a single (production) audience in our account
 # MAILCHIMP_API_KEY=<INTENTIONALLY LEFT BLANK>
 
-# Contentful Config
+# Contentful Config - dedicated development tokens, regenerate any time
 CONTENTFUL_SPACE_ID=vnqe2w0ovn05
 CONTENTFUL_DELIVERY_TOKEN=X2EXmyOCDbKzxVh79tjSlYm7Q5n9FsKFCzeaM8lTsmU
 CONTENTFUL_PREVIEW_TOKEN=hClICMg9L7461N4wfw4G5dawfY8qzUCz8DOj_xiBjD4
